refactor(vc): consolidate note handling back into the vc domain
The standalone note shortcut domain added an extra routing hop for
agents (vc +notes -> note +detail / note +transcript) while the
ownership boundary between vc, note, and minutes was still unsettled,
and spread transcript routing guidance across five skills. Fold note
detail handling back into vc +notes and restore the previous skill
routing until the unified-note design converges.
