+++ cmd/auth/login_test.go
@@ -9,7 +9,6 @@ import (
 	"errors"
 	"io"
 	"net/http"
-	"slices"
 	"sort"
 	"strings"
 	"testing"
@@ -212,12 +211,6 @@ func TestGetShortcutOnlyDomainNames_HaveDescriptions(t *testing.T) {
 		if enDesc == "" {
 			t.Errorf("missing en description for shortcut-only domain %q", name)
 		}
-	}
-}
-
-func TestGetShortcutOnlyDomainNames_IncludesNote(t *testing.T) {
-	if !slices.Contains(getShortcutOnlyDomainNames(), "note") {
-		t.Fatal("shortcut-only domains must include note so auth login can select vc:note:read")
 	}
 }
